+++ src/Aks/Aks.Test/ScenarioTests/KubernetesTests.cs
@@ -100,5 +100,19 @@ namespace Commands.Aks.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-NodeTaints");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestEnableEncryptionAtHost()
+        {
+            TestRunner.RunTestScript("Test-EnableEncryptionAtHost");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestEnableUltraSSD()
+        {
+            TestRunner.RunTestScript("Test-EnableUltraSSD");
+        }
     }
 }